rust/Database: Don't return Option from list()
The only case where this function could return None is if the response
did not contains rows, which can not happen with SELECT requests. If
this happens, it means there is a bug in the driver, Scylla or this
code. This is condition that deserves an error or a panic, not an
Option.