rust: Comment out "song_counter" table creation
This table is not used anywhere. It is also present in C# example, but
unused there too. The only example where it is present in Ruby. That
suggests to me this Rust example is not finished and the table may be
used in the future. For that reason I did not remove its creation, only
commented it out.

Why even touch it at all? It uses "counter" column type, which is not
yet compatible with Tablets, preventing running this example with new
Scylla version.