rust: Don't use Database for migration
Next commit will fix an anti-pattern used in the code: multiple
preparation. Doing this requires preparing statements when creating
Database object, so the migration must be already done then.

The solution is to first create Session, use it for migration, and make
the Database from this session (and prepare the statements then).