rust: Avoid calling deprecated functions
The functions used in this implementation were deprecated in chrono.
Another minor issue is that the implementation was panicking, which is
not a good idea for a function that deserializes potentially
user-provided data.

This commit rewrites the implementation to use new chrono API, and to
avoid panicking.

As a side not, I have no idea why those serde implementations exist here
at all. It looks like a start of some unfinished work, so I'll leave
them be.